Yousef

+++ BE/src/middleware/uploadImage.js
@@ -1,10 +1,10 @@
 import cloudinary from "../utils/cloudinary.js";
 import { asyncHandler } from "../utils/errorHandling.js";
 
-export const uploadImage = (model, isArray = false) => {
-  return asyncHandler(async (req, res, next) => {
+export const uploadImage = ({ model, isArray = false, isFields = false }) => {
+  return async (req, res, next) => {
     const modelName = model.modelName.toLowerCase();
-    if (req.file || req.files?.length) {
+    if (req.file || req.files?.length || Object.keys(req.files).length) {
       if (isArray) {
         const images = [];
 
@@ -23,6 +23,24 @@ export const uploadImage = (model, isArray = false) => {
           });
         }
         req[modelName].images = images;
+      } else if (isFields) {
+        const fields = Object.keys(req.files); // ["images","imageCover"]
+
+        for (const field of fields) {
+          req[modelName][field] = [];
+          for (const obj of req.files[field]) {
+            // upload new imgs
+            const { public_id, secure_url } = await cloudinary.uploader.upload(
+              obj.path,
+              {
+                folder: `web-project-ecommerce/${modelName}/${req[modelName]._id}/${field}`,
+              }
+            );
+
+            req[modelName][field].push({ public_id, secure_url });
+          }
+        }
+        req[modelName].imageCover = req[modelName].imageCover[0];
       } else {
         const { public_id, secure_url } = await cloudinary.uploader.upload(
           req.file.path,
@@ -40,11 +58,11 @@ export const uploadImage = (model, isArray = false) => {
       await req[modelName].save();
     }
 
-    return res.status(200).json({ message: "Done", model: req[modelName] });
-  });
+    return res.status(200).json({ message: "done", model: req[modelName] });
+  };
 };
 
-export const updateImage = (model, isArray = false) => {
+export const updateImage = ({model, isArray = false, isFields = false}) => {
   return asyncHandler(async (req, res, next) => {
     // check if file is send
     if (!(req.file || req.files?.length)) return next();
@@ -76,6 +94,31 @@ export const updateImage = (model, isArray = false) => {
       }
 
       req.body.images = images;
+    } else if (isFields) {
+      const fields = Object.keys(req.files); // ["images","imageCover"]
+
+      for (const field of fields) {
+        req.body[field] = [];
+        for (const obj of req.files[field]) {
+          // delete the pervious images
+          if (modelData[field].length) {
+            await cloudinary.api.delete_resources_by_prefix(
+              `web-project-ecommerce/${modelName}/${modelData._id}/${field}`
+            );
+          }
+
+          // upload new imgs
+          const { public_id, secure_url } = await cloudinary.uploader.upload(
+            obj.path,
+            {
+              folder: `web-project-ecommerce/${modelName}/${modelData._id}/${field}`,
+            }
+          );
+
+          req.body[field].push({ public_id, secure_url });
+        }
+      }
+      if (req.files.imageCover) req.body.imageCover = req.body.imageCover[0];
     } else {
       //delete pervious img
       if (modelData.image)
@@ -101,7 +144,7 @@ export const deleteImage = (model, isArray = false) => {
     const modelData = req[modelName];
 
     // if the model doesn't have image
-    if (!(modelData.image || modelData.images)) return next();
+    if (!(modelData.image || modelData.images || modelData.imageCover)) return next();
 
     if (isArray) {
       await cloudinary.api.delete_resources_by_prefix(
@@ -109,7 +152,11 @@ export const deleteImage = (model, isArray = false) => {
       );
       modelData.images = [];
       await modelData.save();
-    } else {
+    }
+    else if (isFields){
+
+    }
+    else {
       await cloudinary.uploader.destroy(modelData.image.public_id);
       await modelData.updateOne({
         $unset: { image: "" },
